Top-level knit command
The end-user is expected to use knit rather than subcommands directly.
This lets us consistently establish common execution context, like the
KNIT_DIR environment variable. This also enables running knit from any
working directory.

+++ knit-complete-job.sh
@@ -1,7 +1,5 @@
 #!/bin/bash -e
 
-KNIT_DIR=${KNIT_DIR-.knit}
-
 [[ $# -eq 2 ]]
 job_id="$1"
 production_id="$2"
@@ -10,6 +8,7 @@ cache="$KNIT_DIR/jobcache/${job_id:0:2}/${job_id:2}"
 mkdir -p "$(dirname "$cache")"
 exec 3> "$cache.lock"
 flock 3
+trap 'rm "$cache.lock"' EXIT
 
 if [[ -e $cache ]]; then
     echo "Job $job_id has already been completed with production $(<$cache)" >&2
@@ -19,11 +18,10 @@ fi
 echo "$production_id" > $cache.tmp
 mv $cache.tmp $cache
 
+[[ -e $cache.pending ]] || exit
 # TODO possible to orphan pending sessions if we crash
 # TODO drop the lock?
 while IFS=$'\t' read -r session step; do
-    ./record-step-production "$session" "$step" "$production_id"
+    knit-record-step-production "$session" "$step" "$production_id"
 done < $cache.pending
 rm $cache.pending
-
-rm "$cache.lock"
